Split AGRange into its own file.
